+++ transport/tuic/client.go
@@ -4,7 +4,10 @@ import (
 	"context"
 	"crypto/tls"
 	"io"
+<<<<<<< HEAD
 	"math/rand"
+=======
+>>>>>>> 0762b71852a168005a3f133e42dc095278fc607a
 	"net"
 	"os"
 	"runtime"
@@ -58,7 +61,11 @@ func NewClient(options ClientOptions) (*Client, error) {
 		options.Heartbeat = 10 * time.Second
 	}
 	quicConfig := &quic.Config{
+<<<<<<< HEAD
 		DisablePathMTUDiscovery: !(runtime.GOOS == "windows" || runtime.GOOS == "linux" || runtime.GOOS == "android"),
+=======
+		DisablePathMTUDiscovery: !(runtime.GOOS == "windows" || runtime.GOOS == "linux" || runtime.GOOS == "android" || runtime.GOOS == "darwin"),
+>>>>>>> 0762b71852a168005a3f133e42dc095278fc607a
 		MaxDatagramFrameSize:    1400,
 		EnableDatagrams:         true,
 	}
@@ -195,6 +202,7 @@ func (c *Client) ListenPacket(ctx context.Context) (net.PacketConn, error) {
 	if err != nil {
 		return nil, err
 	}
+<<<<<<< HEAD
 	clientPacketConn := newUDPPacketConn(ctx, conn.quicConn, c.udpStream, false)
 	conn.udpAccess.Lock()
 	for {
@@ -206,6 +214,20 @@ func (c *Client) ListenPacket(ctx context.Context) (net.PacketConn, error) {
 		}
 	}
 	conn.udpAccess.Unlock()
+=======
+	var sessionID uint16
+	clientPacketConn := newUDPPacketConn(ctx, conn.quicConn, c.udpStream, false, func() {
+		conn.udpAccess.Lock()
+		delete(conn.udpConnMap, sessionID)
+		conn.udpAccess.Unlock()
+	})
+	conn.udpAccess.Lock()
+	sessionID = conn.udpSessionID
+	conn.udpSessionID++
+	conn.udpConnMap[sessionID] = clientPacketConn
+	conn.udpAccess.Unlock()
+	clientPacketConn.sessionID = sessionID
+>>>>>>> 0762b71852a168005a3f133e42dc095278fc607a
 	return clientPacketConn, nil
 }
 
@@ -218,6 +240,7 @@ func (c *Client) CloseWithError(err error) error {
 }
 
 type clientQUICConnection struct {
+<<<<<<< HEAD
 	quicConn   quic.Connection
 	rawConn    io.Closer
 	access     sync.Mutex
@@ -225,6 +248,16 @@ type clientQUICConnection struct {
 	connErr    error
 	udpAccess  sync.RWMutex
 	udpConnMap map[uint16]*udpPacketConn
+=======
+	quicConn     quic.Connection
+	rawConn      io.Closer
+	closeOnce    sync.Once
+	connDone     chan struct{}
+	connErr      error
+	udpAccess    sync.RWMutex
+	udpConnMap   map[uint16]*udpPacketConn
+	udpSessionID uint16
+>>>>>>> 0762b71852a168005a3f133e42dc095278fc607a
 }
 
 func (c *clientQUICConnection) active() bool {
@@ -233,8 +266,11 @@ func (c *clientQUICConnection) active() bool {
 		return false
 	default:
 	}
+<<<<<<< HEAD
 	c.access.Lock()
 	defer c.access.Unlock()
+=======
+>>>>>>> 0762b71852a168005a3f133e42dc095278fc607a
 	select {
 	case <-c.connDone:
 		return false
@@ -243,6 +279,7 @@ func (c *clientQUICConnection) active() bool {
 	return true
 }
 
+<<<<<<< HEAD
 func (c *clientQUICConnection) handleUDPMessage(message *udpMessage) {
 	c.udpAccess.RLock()
 	udpConn, loaded := c.udpConnMap[message.sessionID]
@@ -271,6 +308,14 @@ func (c *clientQUICConnection) closeWithError(err error) {
 	c.connErr = err
 	close(c.connDone)
 	_ = c.quicConn.CloseWithError(0, "")
+=======
+func (c *clientQUICConnection) closeWithError(err error) {
+	c.closeOnce.Do(func() {
+		c.connErr = err
+		close(c.connDone)
+		_ = c.quicConn.CloseWithError(0, "")
+	})
+>>>>>>> 0762b71852a168005a3f133e42dc095278fc607a
 }
 
 type clientConn struct {