+++ transport/tuic/congestion.go
@@ -38,7 +38,11 @@ func setCongestion(ctx context.Context, connection quic.Connection, congestionNa
 			congestion.NewBBRSender(
 				congestion.DefaultClock{},
 				congestion.GetInitialPacketSize(connection.RemoteAddr()),
+<<<<<<< HEAD
 				10*congestion.InitialMaxDatagramSize,
+=======
+				congestion.InitialCongestionWindow*congestion.InitialMaxDatagramSize,
+>>>>>>> 0762b71852a168005a3f133e42dc095278fc607a
 				congestion.DefaultBBRMaxCongestionWindow*congestion.InitialMaxDatagramSize,
 			),
 		)